northstar-core/requirements.txt
boto3>=1.35
pydantic>=2.0
python-dotenv>=1.0
PyYAML>=6.0
google-cloud-aiplatform>=1.70.0
langgraph
crewai
pyautogen
